DE-1412 Move ValidateEmail() to MailgunImpl (#383)
